+++ src/main/java/wooteco/subway/exception/validation/SectionDistanceExceedException.java
@@ -6,7 +6,7 @@ public class SectionDistanceExceedException extends SubwayValidationException {
 
     private static final String DEFAULT_MESSAGE = "거리 초과로 인해 구간 추가에 실패하였습니다 : ";
 
-    public SectionDistanceExceedException(int distance) {
+    public SectionDistanceExceedException(long distance) {
         super(DEFAULT_MESSAGE + distance);
     }
 }
